@@ -16,24 +16,25 @@
     "node": ">=16.17.0"
   },
   "dependencies": {
+    "@mui/styles": "5.11.16",
     "date-fns": "^2.23.0",
     "deepar": "^4.0.3",
     "focus-trap-react": "^8.7.1",
     "i18next": "^19.8.4",
     "i18next-browser-languagedetector": "^6.0.1",
     "lodash": "^4.17.21",
-    "ra-compact-ui": "^1.1.1",
-    "ra-data-django-rest-framework": "^0.2.0",
-    "ra-i18n-polyglot": "^3.11.1",
-    "ra-language-english": "^3.11.1",
-    "ra-language-french": "^3.11.1",
-    "react": "17.0.2",
-    "react-admin": "^3.19.0",
-    "react-dom": "npm:@hot-loader/react-dom",
+    "ra-compact-ui": "2.0.0-beta",
+    "ra-data-django-rest-framework": "git://github.com/navivanuva/ra-data-django-rest-framework-v4.git",
+    "ra-i18n-polyglot": "^4.9.1",
+    "ra-language-english": "^4.9.1",
+    "ra-language-french": "^4.9.1",
+    "react": "18.2.0",
+    "react-admin": "^4.9.1",
+    "react-dom": "18.2.0",
     "react-hook-form": "^7.12.2",
     "react-hot-loader": "4.12.19",
     "react-i18next": "^11.8.5",
-    "react-router-dom": "^5.2.0",
+    "react-router-dom": "^6.10.0",
     "recharts": "^2.1.6",
     "styled-components": "5.3.0",
     "styled-normalize": "^8.0.7",
@@ -51,8 +52,6 @@
     "@babel/preset-react": "7.8.3",
     "@babel/register": "7.8.6",
     "@babel/runtime": "7.8.4",
-    "@material-ui/data-grid": "^4.0.0-alpha.37",
-    "@material-ui/lab": "^4.0.0-alpha.60",
     "@svgr/webpack": "^5.5.0",
     "@trivago/prettier-plugin-sort-imports": "^2.0.2",
     "babel-core": "7.0.0-bridge.0",
