@@ -19,28 +19,25 @@
  * You should have received a copy of the GNU General Public License
  * along with MEPP.  If not, see <http://www.gnu.org/licenses/>.
  */
-
 import React from 'react';
-import { Admin, Resource } from 'react-admin';
-import { Redirect } from 'react-router';
+import { Admin, CustomRoutes, Resource } from 'react-admin';
+import { Navigate } from 'react-router';
 import { Route } from 'react-router-dom';
 
-import UserIcon from '@material-ui/icons/People';
+import UserIcon from '@mui/icons-material/People';
 
 import Help from '@pages/Help';
 import Intro from '@pages/Intro';
 import Login from '@pages/Login';
 import Mirror from '@pages/Mirror';
+import PlayerStandalonePage from '@pages/PlayerStandAlone';
 import Privacy from '@pages/Privacy';
 import ResetPassword from '@pages/ResetPassword';
 import Terms from '@pages/Terms';
-import PlayerStandalonePage from '@pages/PlayerStandAlone';
 
 import withPage from '@hocs/withPage';
 
-import {
-  CategoryList,
-} from '@components/admin/categories';
+import { CategoryList } from '@components/admin/categories';
 import {
   ClinicianCreate,
   ClinicianEdit,
@@ -67,19 +64,17 @@ import {
   PlanShow,
 } from '@components/admin/plans';
 import { MeppAdminLayout } from '@components/admin/shared';
-import { MeppTheme } from '@components/admin/shared/styles/theme';
-
-import AuthProvider from './authProvider';
-import DataProvider from './dataProvider';
-import i18nProvider from './i18nProvider';
-
 import {
   CategoryIcon,
   ClinicianIcon,
   ExerciseIcon,
   TreatmentPlanIcon,
 } from '@components/admin/shared/icons';
+import { MeppTheme } from '@components/admin/shared/styles/theme';
 
+import AuthProvider from './authProvider';
+import DataProvider from './dataProvider';
+import i18nProvider from './i18nProvider';
 
 export default () => {
   // with auth pages
@@ -94,7 +89,6 @@ export default () => {
   const ResetPasswordPage = withPage(ResetPassword);
   const PlayerAlone = withPage(PlayerStandalonePage);
 
-
   return (
     <Admin
       theme={MeppTheme}
@@ -104,109 +98,75 @@ export default () => {
       authProvider={AuthProvider}
       i18nProvider={i18nProvider}
       loginPage={LoginPage}
-      customRoutes={[
-        <Route
-          exact
-          path="/privacy"
-          component={(props) => <PrivacyPage {...props} />}
-          noLayout
-        />,
-        <Route
-          exact
-          path="/termsofuse"
-          component={(props) => <TermsPage {...props} />}
-          noLayout
-        />,
-        <Route
-          exact
-          path="/support"
-          component={(props) => <HelpPage {...props} />}
-          noLayout
-        />,
-        <Route
-          exact
-          path="/reset-password"
-          component={(props) => <ResetPasswordPage {...props} />}
-          noLayout
-        />,
-        <Route
-          exact
-          path="/intro"
-          component={(props) => <IntroPage {...props} />}
-          noLayout
-        />,
-        <Route
-          exact
-          path="/mirror"
-          component={(props) => <MirrorPage {...props} />}
-          noLayout
-        />,
-        ... process.env.ENVIRONMENT !== 'production'
-          ? [<Route
-              exact
-              path="/playerstandalone"
-              component={(props) => <PlayerAlone {...props} />}
-              noLayout
-           />]
-          : []
-      ]}
       dashboard={Dashboard}
     >
-      {(permissions) => [
-        // Admin panel for users
-        ['admin', 'staff'].includes(permissions) && (
-          <Resource
-            name="patients"
-            create={PatientCreate}
-            edit={PatientEdit}
-            list={PatientList}
-            show={PatientShow}
-            icon={UserIcon}
-          />
-        ),
-        // Admin panel for exercises
-        ['admin', 'staff'].includes(permissions) && (
-          <Resource
-            name="exercises"
-            create={ExerciseCreate}
-            edit={ExerciseEdit}
-            list={ExerciseList}
-            show={ExerciseShow}
-            icon={ExerciseIcon}
-          />
-        ),
-        // Admin panel for treatment plans
-        ['admin', 'staff'].includes(permissions) && (
-          <Resource
-            name="plans"
-            create={PlanCreate}
-            edit={PlanEdit}
-            list={PlanList}
-            show={PlanShow}
-            icon={TreatmentPlanIcon} />
-        ),
-        // Admin panel for clinicians
-        ['admin', 'staff'].includes(permissions) && (
-          <Resource
-            name="clinicians"
-            create={permissions === 'admin' && ClinicianCreate}
-            edit={ClinicianEdit}
-            list={permissions === 'admin' && ClinicianList}
-            show={ClinicianShow}
-            icon={ClinicianIcon}
-          />
-        ),
-        // Admin panel for categories
-        permissions !== 'user' && (
-          <Resource
-            name="categories"
-            list={CategoryList}
-            icon={CategoryIcon}
-          />
-        ),
-        // Mirror app for patients
-        permissions === 'user' && <Redirect to="/intro" />,
-      ]}
+      <CustomRoutes noLayout>
+        <Route exact path="/privacy" element={<PrivacyPage />} />
+        <Route exact path="/termsofuse" element={<TermsPage />} />
+        <Route exact path="/support" element={<HelpPage />} />
+        <Route exact path="/reset-password" element={<ResetPasswordPage />} />
+        <Route exact path="/intro" element={<IntroPage />} />
+        <Route exact path="/mirror" element={<MirrorPage />} />
+        {process.env.ENVIRONMENT !== 'production' ? (
+          <Route exact path="/playerstandalone" component={<PlayerAlone />} />
+        ) : null}
+      </CustomRoutes>
+      {(permissions) => (
+        <>
+          {['admin', 'staff'].includes(permissions) && (
+            <Resource
+              name="patients"
+              create={PatientCreate}
+              edit={PatientEdit}
+              list={PatientList}
+              show={PatientShow}
+              icon={UserIcon}
+            />
+          )}
+          ,
+          {['admin', 'staff'].includes(permissions) && (
+            <Resource
+              name="exercises"
+              create={ExerciseCreate}
+              edit={ExerciseEdit}
+              list={ExerciseList}
+              show={ExerciseShow}
+              icon={ExerciseIcon}
+            />
+          )}
+          ,
+          {['admin', 'staff'].includes(permissions) && (
+            <Resource
+              name="plans"
+              create={PlanCreate}
+              edit={PlanEdit}
+              list={PlanList}
+              show={PlanShow}
+              icon={TreatmentPlanIcon}
+            />
+          )}
+          ,
+          {['admin', 'staff'].includes(permissions) && (
+            <Resource
+              name="clinicians"
+              create={permissions === 'admin' && ClinicianCreate}
+              edit={ClinicianEdit}
+              list={permissions === 'admin' && ClinicianList}
+              show={ClinicianShow}
+              icon={ClinicianIcon}
+            />
+          )}
+          ,
+          {permissions !== 'user' && (
+            <Resource
+              name="categories"
+              list={CategoryList}
+              icon={CategoryIcon}
+            />
+          )}
+          ,{permissions === 'user' && <Navigate to="/intro" />}
+        </>
+      )}
     </Admin>
   );
 };
