@@ -19,7 +19,6 @@
  * You should have received a copy of the GNU General Public License
  * along with MEPP.  If not, see <http://www.gnu.org/licenses/>.
  */
-
 import React from 'react';
 import {
   List,
@@ -28,21 +27,25 @@ import {
   useLocale,
   useTranslate,
 } from 'react-admin';
-import { Chip } from '@material-ui/core';
-import { useTheme } from '@material-ui/core/styles';
+
+import { Chip } from '@mui/material';
+import { useTheme } from '@mui/material/styles';
 
 const SubCategoriesRow = ({ record, locale }) => {
   const theme = useTheme();
   return (
-    <div style={{display: 'inline-flex', flexWrap: 'wrap', gap: theme.spacing(1)}}>
+    <div
+      style={{
+        display: 'inline-flex',
+        flexWrap: 'wrap',
+        gap: theme.spacing(1),
+      }}
+    >
       {record.sub_categories.map((subCategory) => (
-        <Chip
-          key={subCategory.id}
-          label={subCategory.i18n.name[locale]}
-        />
+        <Chip key={subCategory.id} label={subCategory.i18n.name[locale]} />
       ))}
     </div>
-  )
+  );
 };
 
 export const CategoryList = (props) => {
@@ -57,10 +60,12 @@ export const CategoryList = (props) => {
       filterDefaultValues={{
         language: locale,
       }}
-      bulkActionButtons={false}
       actions={false}
     >
-      <Datagrid expand={<SubCategoriesRow locale={locale}/>}>
+      <Datagrid
+        bulkActionButtons={false}
+        expand={<SubCategoriesRow locale={locale} />}
+      >
         <TextField source={`i18n.name.${locale}`} />
       </Datagrid>
     </List>
