@@ -19,8 +19,8 @@
  * You should have received a copy of the GNU General Public License
  * along with MEPP.  If not, see <http://www.gnu.org/licenses/>.
  */
-
-import React, {Fragment, useMemo, useState} from 'react';
+import { fetchJsonWithAuthToken } from 'ra-data-django-rest-framework';
+import React, { Fragment, useMemo, useState } from 'react';
 import {
   useGetIdentity,
   useGetList,
@@ -28,6 +28,9 @@ import {
   useNotify,
   useTranslate,
 } from 'react-admin';
+
+import CheckCircle from '@mui/icons-material/CheckCircle';
+import ErrorOutlineIcon from '@mui/icons-material/ErrorOutline';
 import {
   Box,
   Card,
@@ -37,19 +40,18 @@ import {
   DialogContent,
   TextField as TextFieldMui,
   CircularProgress,
-  DialogActions, Dialog,
-} from '@material-ui/core';
-import { makeStyles } from '@material-ui/core/styles';
-import { MirrorIcon } from '@components/admin/shared/icons/MirrorIcon';
-import { fetchJsonWithAuthToken } from 'ra-data-django-rest-framework';
-import Autocomplete from '@material-ui/lab/Autocomplete';
-import ErrorOutlineIcon from '@material-ui/icons/ErrorOutline';
-import CheckCircle from '@material-ui/icons/CheckCircle';
+  DialogActions,
+  Dialog,
+} from '@mui/material';
+import Autocomplete from '@mui/material/Autocomplete';
+import { makeStyles } from '@mui/styles';
 
-const useStyles = makeStyles(theme => ({
+import { MirrorIcon } from '@components/admin/shared/icons/MirrorIcon';
+
+const useStyles = makeStyles((theme) => ({
   actions: {
     justifyContent: 'center',
-    [theme.breakpoints.down('sm')]: {
+    [theme.breakpoints.down('md')]: {
       marginTop: 0,
       padding: 0,
       flexWrap: 'wrap',
@@ -67,12 +69,16 @@ export const Welcome = () => {
   const t = useTranslate();
   const locale = useLocale();
   const notify = useNotify();
-  const { loading: identityLoading } = useGetIdentity();
+  const { isLoading: identityLoading } = useGetIdentity();
   const [confirmDisabled, setConfirmDisabled] = useState(true);
   const [openDialog, setOpenDialog] = useState(false);
   const [selectedPatient, setSelectedPatient] = useState(undefined);
-  const handleOpenDialog = () => { setOpenDialog(true) };
-  const handleCloseDialog = () => { setOpenDialog(false); };
+  const handleOpenDialog = () => {
+    setOpenDialog(true);
+  };
+  const handleCloseDialog = () => {
+    setOpenDialog(false);
+  };
   const handleAutocompleteChange = (event, patient) => {
     setSelectedPatient(patient);
     setConfirmDisabled(patient ? false : true);
@@ -87,7 +93,9 @@ export const Welcome = () => {
       body: `{"type":"mirror", "patient_uid": "${selectedPatient.id}"}`,
     })
       .then((response) => {
-        const qs = new URLSearchParams({tt: response.json['token']}).toString();
+        const qs = new URLSearchParams({
+          tt: response.json['token'],
+        }).toString();
         const mirrorUrl = `/?${qs}#/intro`;
         setOpenDialog(false);
         setSelectedPatient(undefined);
@@ -98,22 +106,25 @@ export const Welcome = () => {
         notify('admin.shared.notifications.mirror.failure', 'error');
       });
   };
-  const {data, loading: patientsLoading, loaded} = useGetList(
-    'patients',
-    { page: 1, perPage: 9999},
-    { field: 'full_name', order: 'ASC' },
-    {
+  const { data, isLoading: patientsLoading } = useGetList('patients', {
+    pagination: { page: 1, perPage: 9999 },
+    sort: { field: 'full_name', order: 'ASC' },
+    filter: {
       language: locale,
-      archived: false
-    }
-  );
+      archived: false,
+    },
+  });
 
   const patients = useMemo(() => {
-    return Object.values(data).map((patient) => ({
-      name: patient.full_name,
-      id: patient.id
-    }));
-  }, [data, loaded]);
+    if (data) {
+      return Object.values(data).map((patient) => ({
+        name: patient.full_name,
+        id: patient.id,
+      }));
+    } else {
+      return [];
+    }
+  }, [data]);
 
   if (identityLoading) return <></>;
 
@@ -130,10 +141,10 @@ export const Welcome = () => {
             >
               {t('admin.dashboard.labels.openMirrorButton')}
             </Button>
-            <Dialog
-              open={openDialog}
-            >
-              <DialogTitle>{t('admin.dashboard.mirror_dialog.title')}</DialogTitle>
+            <Dialog open={openDialog}>
+              <DialogTitle>
+                {t('admin.dashboard.mirror_dialog.title')}
+              </DialogTitle>
               <DialogContent>
                 <Autocomplete
                   options={patients}
@@ -142,14 +153,19 @@ export const Welcome = () => {
                   onChange={handleAutocompleteChange}
                   getOptionLabel={(option) => option.name}
                   renderInput={(params) => (
-                    <TextFieldMui {...params}
-                      label={t('admin.dashboard.mirror_dialog.labels.autocomplete')}
+                    <TextFieldMui
+                      {...params}
+                      label={t(
+                        'admin.dashboard.mirror_dialog.labels.autocomplete',
+                      )}
                       variant="filled"
                       InputProps={{
                         ...params.InputProps,
                         endAdornment: (
                           <Fragment>
-                            {patientsLoading ? <CircularProgress color="inherit" size={20} /> : null}
+                            {patientsLoading ? (
+                              <CircularProgress color="inherit" size={20} />
+                            ) : null}
                             {params.InputProps.endAdornment}
                           </Fragment>
                         ),
