@@ -19,8 +19,7 @@
  * You should have received a copy of the GNU General Public License
  * along with MEPP.  If not, see <http://www.gnu.org/licenses/>.
  */
-
-import React, { Fragment, useEffect, useState }  from 'react';
+import React, { Fragment, useEffect, useState } from 'react';
 import {
   Datagrid,
   FunctionField,
@@ -30,28 +29,35 @@ import {
   TextField,
   useListContext,
   useLocale,
+  usePermissions,
+  useResourceContext,
   useTranslate,
 } from 'react-admin';
-import { Divider, Tabs, Tab } from '@material-ui/core';
+
+import { Divider, Tabs, Tab } from '@mui/material';
+
+import PlanListAside from '@components/admin/plans/PlanListAside';
+import BulkActionButtons from '@components/admin/shared/toolbars/BulkActionsToolbar';
+import ListActions from '@components/admin/shared/toolbars/ListToolbar';
 
 import Spinner from '../shared/Spinner';
 import ArchivableFilter from '../shared/filters/ArchivableFilter';
 import RowActionToolbar from '../shared/toolbars/RowActionToolbar';
-import ListActions from '@components/admin/shared/toolbars/ListToolbar';
-import BulkActionButtons from '@components/admin/shared/toolbars/BulkActionsToolbar';
-import PlanListAside from '@components/admin/plans/PlanListAside';
 
 const tabs = [
   { id: 'user', is_system: false },
   { id: 'system', is_system: true },
 ];
 
-const PlanDatagrid = ({permissions, ...props}) => {
+const PlanDatagrid = ({ permissions, ...props }) => {
   const locale = useLocale();
   const t = useTranslate();
 
   return (
-    <Datagrid {...props}>
+    <Datagrid
+      {...props}
+      bulkActionButtons={<BulkActionButtons permissions={permissions} />}
+    >
       <TextField source={`i18n.name.${locale}`} />
       <TextField textAlign="center" source="daily_repeat" />
       <FunctionField
@@ -60,26 +66,21 @@ const PlanDatagrid = ({permissions, ...props}) => {
         render={(record) => record.exercises.length}
       />
       {permissions === 'admin' && (
-        <ReferenceField
-          source="clinician_uid"
-          reference="clinicians"
-        >
+        <ReferenceField source="clinician_uid" reference="clinicians">
           <TextField source="full_name" />
         </ReferenceField>
       )}
-      <RowActionToolbar
-        permissions={permissions}
-        clonable={true}
-      />
+      <RowActionToolbar permissions={permissions} clonable={true} />
     </Datagrid>
   );
 };
 
-const TabbedDatagrid = ({permissions, ...props}) => {
+const TabbedDatagrid = ({ permissions, ...props }) => {
   const listContext = useListContext();
   const t = useTranslate();
   const { ids, filterValues, setFilters, displayedFilters, loading } =
     listContext;
+  const resource = useResourceContext();
   const [userPlanIds, setUserPlanIds] = useState([]);
   const [systemPlanIds, setSystemPlanIds] = useState([]);
   const [value, setValue] = useState('user');
@@ -95,7 +96,9 @@ const TabbedDatagrid = ({permissions, ...props}) => {
   useEffect(() => {
     if (filterValues.is_system) {
       // reset `value` to `system` when it's out of sync
-      if (value === 'user') { setValue('system'); }
+      if (value === 'user') {
+        setValue('system');
+      }
       setSystemPlanIds(ids);
     } else {
       setUserPlanIds(ids);
@@ -104,22 +107,13 @@ const TabbedDatagrid = ({permissions, ...props}) => {
 
   return (
     <Fragment>
-      <Tabs
-        value={value}
-        indicatorColor="primary"
-        onChange={handleChange}
-      >
+      <Tabs value={value} indicatorColor="primary" onChange={handleChange}>
         {tabs.map((choice) => {
-          const label = permissions === 'admin' && choice.id === 'user'
-            ? t(`resources.${props.resource}.list.labels.admin`)
-            : t(`resources.${props.resource}.list.labels.${choice.id}`);
-          return (
-            <Tab
-              key={choice.id}
-              label={label}
-              value={choice.id}
-            />
-          )
+          const label =
+            permissions === 'admin' && choice.id === 'user'
+              ? t(`resources.${resource}.list.labels.admin`)
+              : t(`resources.${resource}.list.labels.${choice.id}`);
+          return <Tab key={choice.id} label={label} value={choice.id} />;
         })}
       </Tabs>
       <Divider />
@@ -141,6 +135,7 @@ const TabbedDatagrid = ({permissions, ...props}) => {
 };
 
 export const PlanList = (props) => {
+  const { permissions } = usePermissions();
   const locale = useLocale();
   return (
     <List
@@ -154,11 +149,10 @@ export const PlanList = (props) => {
       filters={<ArchivableFilter />}
       sort={{ field: `i18n.description.${locale}`, order: 'ASC' }}
       perPage={25}
-      bulkActionButtons={<BulkActionButtons permissions={props.permissions} />}
-      actions={<ListActions/>}
-      aside={<PlanListAside permissions={props.permissions} />}
+      actions={<ListActions />}
+      aside={<PlanListAside permissions={permissions} />}
     >
-      <TabbedDatagrid permissions={props.permissions}/>
+      <TabbedDatagrid permissions={permissions} />
     </List>
-  )
+  );
 };
