@@ -25,8 +25,8 @@ import {
   useRedirect,
   useTranslate,
 } from 'react-admin';
-import Button from '@material-ui/core/Button';
-import ArrowBackIosIcon from '@material-ui/icons/ArrowBackIos';
+import Button from '@mui/material/Button';
+import ArrowBackIosIcon from '@mui/icons-material/ArrowBackIos';
 
 const BackButton = ({ patientUid, basePath }) => {
   const redirect = useRedirect();
