@@ -21,7 +21,7 @@
  */
 
 import React from 'react';
-import { Typography as MuiTypography} from '@material-ui/core';
+import { Typography as MuiTypography} from '@mui/material';
 import {sanitizeRestProps} from '@admin/utils/props';
 
 export const Div = (props) => {
