@@ -21,7 +21,7 @@
  */
 
 import React from 'react';
-import SvgIcon from '@material-ui/core/SvgIcon';
+import SvgIcon from '@mui/material/SvgIcon';
 
 const logoStyle = {
   margin: '18px 0 10px 15px',
