@@ -22,20 +22,23 @@
 import { DeepAR } from 'deepar';
 import React, { useRef, useEffect, useContext } from 'react';
 import { useGetIdentity } from 'react-admin';
-import { useHistory } from 'react-router-dom';
+import { useNavigate } from 'react-router-dom';
 import styled from 'styled-components';
 
+
+
 import { ExerciseStep, ExerciseContext } from './ExerciseProvider';
+
 
 /**
  * DeepAP.ai player
  */
 const Player = () => {
   const { exercise, exerciseStep, ready } = useContext(ExerciseContext);
-  const { identity, loading: identityLoading } = useGetIdentity();
+  const { identity, isLoading: identityLoading } = useGetIdentity();
   const deepAR = useRef(null);
   const canvas = useRef(null);
-  const history = useHistory();
+  const navigate = useNavigate();
 
   /**
    * Init Deep AR
@@ -44,7 +47,6 @@ const Player = () => {
   useEffect(() => {
     if (!identity) return;
     if (deepAR.current) return;
-
     canvas.current.width = window.innerWidth;
     canvas.current.height = window.innerHeight;
 
@@ -67,7 +69,7 @@ const Player = () => {
       },
     });
     AR.downloadFaceTrackingModel('./assets/deepar/models-68-extreme.bin');
-    AR.callbacks.onCameraPermissionDenied = () => history.push('/intro');
+    AR.callbacks.onCameraPermissionDenied = () => navigate('/intro');
     deepAR.current = AR;
   }, [deepAR, canvas, identity, exercise]);
 
