@@ -19,22 +19,24 @@
  * You should have received a copy of the GNU General Public License
  * along with MEPP.  If not, see <http://www.gnu.org/licenses/>.
  */
-
+import { RaBox } from 'ra-compact-ui';
 import React from 'react';
-import { Typography } from '@components/admin/shared/dom/sanitize';
-import { makeStyles } from '@material-ui/core/styles';
-import { CompactForm, RaBox } from 'ra-compact-ui';
 import {
   Create,
   SelectInput,
+  SimpleForm,
   PasswordInput,
   TextInput,
   BooleanInput,
+  useResourceContext,
   useTranslate,
   useNotify,
 } from 'react-admin';
 
-import SimpleFormToolBar from '../shared/toolbars/SimpleFormToolBar';
+import { makeStyles } from '@mui/styles';
+
+import { Typography } from '@components/admin/shared/dom/sanitize';
+import Options from '@components/admin/shared/options';
 import {
   validateEmail,
   validateFirstName,
@@ -43,7 +45,8 @@ import {
   validatePasswordRequired as validatePassword,
   validatePasswords,
 } from '@components/admin/shared/validators';
-import Options from '@components/admin/shared/options';
+
+import SimpleFormToolBar from '../shared/toolbars/SimpleFormToolbar';
 
 const useStyles = makeStyles((theme) => ({
   root: {
@@ -56,6 +59,7 @@ const useStyles = makeStyles((theme) => ({
 }));
 
 export const ClinicianCreate = (props) => {
+  const resourceName = useResourceContext();
   const t = useTranslate();
   const classes = useStyles();
   const options = Options();
@@ -63,20 +67,15 @@ export const ClinicianCreate = (props) => {
   const onFailure = (error) => {
     let message = '';
     Object.entries(error.body).forEach(([key, values]) => {
-      message += t(`resources.${props.resource}.errors.${key}`);
+      message += t(`resources.${resourceName}.errors.${key}`);
     });
-    notify(message, {type: 'error'});
+    notify(message, { type: 'error' });
   };
 
   return (
-    <Create
-      onFailure={onFailure}
-      {...props}
-    >
-      <CompactForm
-        layoutComponents={[RaBox]}
-        toolbar={<SimpleFormToolBar identity={false}/>}
-        validate={validatePasswords}
+    <Create queryOptions={{ onError: onFailure }} {...props}>
+      <SimpleForm
+        toolbar={<SimpleFormToolBar identity={false} />}
         redirect="list"
       >
         <Typography variant="h6" gutterBottom>
@@ -88,33 +87,23 @@ export const ClinicianCreate = (props) => {
             fullWidth
             validate={validateFirstName}
           />
-          <TextInput
-            source="last_name"
-            fullWidth
-            validate={validateLastName}
-          />
+          <TextInput source="last_name" fullWidth validate={validateLastName} />
         </RaBox>
         <RaBox className={classes.root}>
-          <TextInput
-            source="email"
-            fullWidth
-            validate={validateEmail}
-          />
+          <TextInput source="email" fullWidth validate={validateEmail} />
         </RaBox>
         <Typography variant="h6" gutterBottom gutterTop={true}>
           {t('admin.shared.labels.card.informations')}
         </Typography>
-          <RaBox className={classes.root}>
-            <SelectInput
-              source="language"
-              choices={options.languages}
-              fullWidth
-              validate={validateLanguage}
-            />
-            <BooleanInput
-              source="is_superuser"
-            />
-          </RaBox>
+        <RaBox className={classes.root}>
+          <SelectInput
+            source="language"
+            choices={options.languages}
+            fullWidth
+            validate={validateLanguage}
+          />
+          <BooleanInput source="is_superuser" />
+        </RaBox>
         <Typography variant="h6" gutterBottom gutterTop={true}>
           {t('admin.shared.labels.card.create_password')}
         </Typography>
@@ -128,10 +117,11 @@ export const ClinicianCreate = (props) => {
           <PasswordInput
             label={t('resources.clinicians.fields.confirm_password')}
             source="confirm_password"
+            validate={validatePasswords}
             fullWidth
           />
         </RaBox>
-      </CompactForm>
+      </SimpleForm>
     </Create>
   );
 };
