@@ -19,23 +19,26 @@
  * You should have received a copy of the GNU General Public License
  * along with MEPP.  If not, see <http://www.gnu.org/licenses/>.
  */
-
+import { RaBox } from 'ra-compact-ui';
 import React from 'react';
-import { Typography } from '@components/admin/shared/dom/sanitize';
-import { makeStyles } from '@material-ui/core/styles';
-import { CompactForm, RaBox } from 'ra-compact-ui';
 import {
   BooleanInput,
   Edit,
   SelectInput,
+  SimpleForm,
   PasswordInput,
   TextInput,
   useGetIdentity,
+  useResourceContext,
   useTranslate,
   useNotify,
 } from 'react-admin';
 
-import SimpleFormToolBar from '../shared/toolbars/SimpleFormToolBar';
+import { makeStyles } from '@mui/styles';
+
+import { Typography } from '@components/admin/shared/dom/sanitize';
+import Options from '@components/admin/shared/options';
+import TopToolbar from '@components/admin/shared/toolbars/TopToolbar';
 import {
   validateEmail,
   validateFirstName,
@@ -44,8 +47,8 @@ import {
   validatePasswordOptional as validatePassword,
   validatePasswords,
 } from '@components/admin/shared/validators';
-import Options from '@components/admin/shared/options';
-import TopToolbar from "@components/admin/shared/toolbars/TopToolbar";
+
+import SimpleFormToolBar from '../shared/toolbars/SimpleFormToolbar';
 
 const useStyles = makeStyles((theme) => ({
   root: {
@@ -57,8 +60,8 @@ const useStyles = makeStyles((theme) => ({
   },
 }));
 
-const ProfileRow = ({identity, identityLoaded, ...props}) => {
-  if (!identityLoaded || identity?.uid === props?.record?.id) {
+const ProfileRow = ({ identity, identityLoading, ...props }) => {
+  if (identityLoading || identity?.uid === props?.record?.id) {
     return false;
   } else {
     return props.children;
@@ -67,29 +70,27 @@ const ProfileRow = ({identity, identityLoaded, ...props}) => {
 
 export const ClinicianEdit = (props) => {
   const t = useTranslate();
-  const { identity, loaded: identityLoaded } = useGetIdentity();
+  const { identity, isLoading: identityLoading } = useGetIdentity();
+  const resourceName = useResourceContext();
   const classes = useStyles();
   const notify = useNotify();
   const options = Options();
   const onFailure = (error) => {
     let message = '';
     Object.entries(error.body).forEach(([key, values]) => {
-      message += t(`resources.${props.resource}.errors.${key}`);
+      message += t(`resources.${resourceName}.errors.${key}`);
     });
-    notify(message, {type: 'error'});
+    notify(message, { type: 'error' });
   };
 
   return (
     <Edit
-      onFailure={onFailure}
-      undoable={false}
+      queryOptions={{ onError: onFailure }}
       actions={<TopToolbar identity={identity} />}
       {...props}
     >
-      <CompactForm
-        layoutComponents={[RaBox]}
+      <SimpleForm
         toolbar={<SimpleFormToolBar identity={identity} />}
-        validate={validatePasswords}
       >
         <Typography variant="h6" gutterBottom>
           {t('admin.shared.labels.card.identity')}
@@ -100,20 +101,12 @@ export const ClinicianEdit = (props) => {
             fullWidth
             validate={validateFirstName}
           />
-          <TextInput
-            source="last_name"
-            fullWidth
-            validate={validateLastName}
-          />
+          <TextInput source="last_name" fullWidth validate={validateLastName} />
         </RaBox>
         <RaBox className={classes.root}>
-          <TextInput
-            source="email"
-            fullWidth
-            validate={validateEmail}
-          />
+          <TextInput source="email" fullWidth validate={validateEmail} />
         </RaBox>
-        <ProfileRow identity={identity} identityLoaded={identityLoaded}>
+        <ProfileRow identity={identity} identityLoading={identityLoading}>
           <Typography variant="h6" gutterBottom gutterTop={true}>
             {t('admin.shared.labels.card.informations')}
           </Typography>
@@ -144,10 +137,11 @@ export const ClinicianEdit = (props) => {
           <PasswordInput
             label={t('resources.clinicians.fields.confirm_password')}
             source="confirm_password"
+            validate={validatePasswords}
             fullWidth
           />
         </RaBox>
-      </CompactForm>
+      </SimpleForm>
     </Edit>
   );
 };
