@@ -19,12 +19,12 @@
  * You should have received a copy of the GNU General Public License
  * along with MEPP.  If not, see <http://www.gnu.org/licenses/>.
  */
-
 import React, { useState } from 'react';
 import {
   ArrayInput,
   BooleanInput,
   Create,
+  FormDataConsumer,
   NumberInput,
   SelectInput,
   SimpleForm,
@@ -33,30 +33,31 @@ import {
   TranslatableInputs,
   useGetList,
   useLocale,
+  usePermissions,
   useTranslate,
 } from 'react-admin';
-import {
-  validateCategory,
-  validateSubCategory,
-  validateSubCategories,
-} from '@components/admin/exercises/validators';
-import { validateNumber } from '@components/admin/shared/validators';
+
+import SubCategoryInput from '@components/admin/exercises/SubCategoryInput';
+import { preSave } from '@components/admin/exercises/callbacks';
 import {
   useNumberStyles,
   useSimpleFormIteratorStyles,
   useTranslatorInputStyles,
 } from '@components/admin/exercises/styles';
-import SubCategoryInput from '@components/admin/exercises/SubCategoryInput';
-import { LANGUAGES } from '../../../locales';
-import {requiredLocalizedField} from '@components/admin/shared/validators';
-import SimpleFormToolBar from '@components/admin/shared/toolbars/SimpleFormToolBar';
 import {
-  preSave
-} from '@components/admin/exercises/callbacks';
+  validateCategory,
+  validateSubCategory,
+  validateSubCategories,
+} from '@components/admin/exercises/validators';
 import { Typography, Div } from '@components/admin/shared/dom/sanitize';
+import SimpleFormToolBar from '@components/admin/shared/toolbars/SimpleFormToolbar';
+import { validateNumber } from '@components/admin/shared/validators';
+import { requiredLocalizedField } from '@components/admin/shared/validators';
+
+import { LANGUAGES } from '../../../locales';
 
 export const ExerciseCreate = (props) => {
-
+  const { permissions } = usePermissions();
   const t = useTranslate();
   const simpleFormIteratorclasses = useSimpleFormIteratorStyles();
   const numberClasses = useNumberStyles();
@@ -65,16 +66,19 @@ export const ExerciseCreate = (props) => {
   const [updatedSubCategoryInputs, setUpdatedSubCategoryInputs] = useState({});
   let categories = [];
   let subCategories = {};
-  const {data, ids, loaded} = useGetList(
+  const { data, isLoading } = useGetList(
     'categories',
-    { page: 1, perPage: 9999},
-    { field: 'i18n__name', order: 'ASC' },
-    { language: locale }
+    {
+      pagination: { page: 1, perPage: 9999 },
+      sort: { field: 'i18n__name', order: 'ASC' },
+    },
+    { language: locale },
   );
 
-  const validateI18n = (record) => {
-    return requiredLocalizedField(record, locale, 'description');
-  }
+  const validateI18n = (value, record) => {
+    return requiredLocalizedField(value, record, locale, 'description');
+  };
+
   /* Update description translations if empty */
   const transform = (record) => {
     return preSave(record, locale);
@@ -88,28 +92,32 @@ export const ExerciseCreate = (props) => {
     updates[categoryInput.name.replace('category__', '')] = categoryInput.value;
     setUpdatedSubCategoryInputs({
       ...updatedSubCategoryInputs,
-      ...updates
+      ...updates,
     });
   };
 
   // ToDo refactor
-  if (loaded) {
-    ids.forEach((categoryUid) => {
-      categories.push({'id': categoryUid, 'name': data[categoryUid].i18n.name[locale]});
-      subCategories[categoryUid] = data[categoryUid]['sub_categories'].map((subCategory) => {
-        return {'id': subCategory.id, 'name': subCategory.i18n.name[locale]};
+  if (!isLoading) {
+    data.forEach((category) => {
+      categories.push({
+        'id': category.id,
+        'name': category.i18n.name[locale],
       });
+      subCategories[category.id] = category.sub_categories.map(
+        (subCategory) => {
+          return {
+            'id': subCategory.id,
+            'name': subCategory.i18n.name[locale],
+          };
+        },
+      );
     });
   }
 
   return (
-    <Create
-      transform={transform}
-      {...props}
-    >
+    <Create transform={transform} {...props}>
       <SimpleForm
         redirect="show"
-        validate={validateI18n}
         toolbar={<SimpleFormToolBar identity={false} />}
       >
         <Typography variant="h6" gutterBottom>
@@ -124,13 +132,10 @@ export const ExerciseCreate = (props) => {
             source="i18n.description"
             multiline={true}
             fullWidth={true}
+            validate={validateI18n}
           />
         </TranslatableInputs>
-        {props.permissions === 'admin' &&
-          <BooleanInput
-            source="is_system"
-          />
-        }
+        {permissions === 'admin' && <BooleanInput source="is_system" />}
         <Div className={numberClasses.numbers}>
           <NumberInput
             source="movement_duration"
@@ -151,7 +156,7 @@ export const ExerciseCreate = (props) => {
         <Typography variant="h6" gutterBottom gutterTop={true}>
           {t('resources.exercises.card.labels.classification')}
         </Typography>
-        {loaded && (
+        {!isLoading && (
           <ArrayInput
             source="sub_categories"
             validate={validateSubCategories}
@@ -162,17 +167,26 @@ export const ExerciseCreate = (props) => {
               disableReordering={true}
             >
               <SelectInput
+                label="Category"
                 source="category__uid"
                 choices={categories}
                 onChange={handleChange}
                 validate={validateCategory}
               />
-              <SubCategoryInput
-                source="uid"
-                updatedSubCategoryInputs={updatedSubCategoryInputs}
-                subCategories={subCategories}
-                validate={validateSubCategory}
-              />
+              <FormDataConsumer>
+                {({ scopedFormData, getSource, ...rest }) =>
+                  scopedFormData ? (
+                    <SubCategoryInput
+                      label="Sub-category"
+                      source={getSource('uid')}
+                      data={scopedFormData}
+                      updatedSubCategoryInputs={updatedSubCategoryInputs}
+                      subCategories={subCategories}
+                      validate={validateSubCategory}
+                    />
+                  ) : null
+                }
+              </FormDataConsumer>
             </SimpleFormIterator>
           </ArrayInput>
         )}
