@@ -19,26 +19,25 @@
  * You should have received a copy of the GNU General Public License
  * along with MEPP.  If not, see <http://www.gnu.org/licenses/>.
  */
-
-import { Typography } from '@components/admin/shared/dom/sanitize';
-import { makeStyles } from '@material-ui/core/styles';
-import { CompactForm, RaBox } from 'ra-compact-ui';
+import { RaBox } from 'ra-compact-ui';
 import React from 'react';
 import {
   Create,
   SelectInput,
   PasswordInput,
   ReferenceInput,
+  SimpleForm,
   TextInput,
+  usePermissions,
+  useResourceContext,
   useTranslate,
   useNotify,
 } from 'react-admin';
-import SimpleFormToolBar from '../shared/toolbars/SimpleFormToolBar';
-import {
-  validateAudio,
-  validateClinician,
-  validateSide,
-} from './validators';
+
+import { makeStyles } from '@mui/styles';
+
+import { Typography } from '@components/admin/shared/dom/sanitize';
+import Options from '@components/admin/shared/options';
 import {
   validateEmail,
   validateFirstName,
@@ -47,7 +46,9 @@ import {
   validatePasswordRequired as validatePassword,
   validatePasswords,
 } from '@components/admin/shared/validators';
-import Options from '@components/admin/shared/options';
+
+import SimpleFormToolBar from '../shared/toolbars/SimpleFormToolbar';
+import { validateAudio, validateClinician, validateSide } from './validators';
 
 const useStyles = makeStyles((theme) => ({
   root: {
@@ -60,28 +61,25 @@ const useStyles = makeStyles((theme) => ({
 }));
 
 export const PatientCreate = (props) => {
+  const resourceName = useResourceContext();
   const t = useTranslate();
   const classes = useStyles();
   const options = Options();
   const notify = useNotify();
-  const onFailure = (error) => {
+  const { permissions } = usePermissions();
+  const handleFailure = (error) => {
     let message = '';
     Object.entries(error.body).forEach(([key, values]) => {
-      message += t(`resources.${props.resource}.errors.${key}`);
+      message += t(`resources.${resourceName}.errors.${key}`);
     });
-    notify(message, {type: 'error'});
+    notify(message, { type: 'error' });
   };
 
   return (
-    <Create
-      onFailure={onFailure}
-      {...props}
-    >
-      <CompactForm
-        layoutComponents={[RaBox]}
+    <Create mutationOptions={{ onError: handleFailure }} {...props}>
+      <SimpleForm
         redirect="show"
-        toolbar={<SimpleFormToolBar identity={false}/>}
-        validate={validatePasswords}
+        toolbar={<SimpleFormToolBar identity={false} />}
       >
         <Typography variant="h6" gutterBottom>
           {t('admin.shared.labels.card.identity')}
@@ -114,7 +112,7 @@ export const PatientCreate = (props) => {
             fullWidth
           />
         </RaBox>
-        {props.permissions === 'admin' && (
+        {permissions === 'admin' && (
           <RaBox className={classes.root}>
             <ReferenceInput
               source="clinician_uid"
@@ -133,7 +131,7 @@ export const PatientCreate = (props) => {
             />
           </RaBox>
         )}
-        {props.permissions !== 'admin' && (
+        {permissions !== 'admin' && (
           <RaBox className={classes.root}>
             <SelectInput
               source="language"
@@ -156,10 +154,11 @@ export const PatientCreate = (props) => {
           <PasswordInput
             label={t('resources.patients.fields.confirm_password')}
             source="confirm_password"
+            validate={validatePasswords}
             fullWidth
           />
         </RaBox>
-      </CompactForm>
+      </SimpleForm>
     </Create>
   );
 };
