@@ -19,20 +19,9 @@
  * You should have received a copy of the GNU General Public License
  * along with MEPP.  If not, see <http://www.gnu.org/licenses/>.
  */
-
-import React, {Fragment, useState} from 'react';
-import {
-  IconButton,
-  FormGroup,
-  FormControlLabel,
-  Switch,
-  DialogTitle,
-  DialogContent,
-  DialogActions, Button, Dialog, DialogContentText
-} from '@material-ui/core';
-import MailOutlineIcon from '@material-ui/icons/MailOutline';
-import { makeStyles } from '@material-ui/core/styles';
 import { BoxedShowLayout, RaBox } from 'ra-compact-ui';
+import { fetchJsonWithAuthToken } from 'ra-data-django-rest-framework';
+import React, { Fragment, useState } from 'react';
 import {
   Datagrid,
   Show,
@@ -45,20 +34,38 @@ import {
   useLocale,
   useGetList,
   useNotify,
+  usePermissions,
+  useRecordContext,
+  useResourceContext,
   useTranslate,
 } from 'react-admin';
 
-import AddPlanButton from '@components/admin/patients/AddPlanButton';
+import CheckCircle from '@mui/icons-material/CheckCircle';
+import ErrorOutlineIcon from '@mui/icons-material/ErrorOutline';
+import MailOutlineIcon from '@mui/icons-material/MailOutline';
+import {
+  IconButton,
+  FormGroup,
+  FormControlLabel,
+  Switch,
+  DialogTitle,
+  DialogContent,
+  DialogActions,
+  Button,
+  Dialog,
+  DialogContentText,
+} from '@mui/material';
+import { makeStyles } from '@mui/styles';
+
+import { sanitizeRestProps } from '@admin/utils/props';
+
 import ClinicianTextField from '@components/admin/clinicians/ClinicianTextField';
-import ShowToolBar from '@components/admin/shared/toolbars/ShowToolbar';
+import AddPlanButton from '@components/admin/patients/AddPlanButton';
 import Spinner from '@components/admin/shared/Spinner';
-import RowActionToolbar from '@components/admin/shared/toolbars/RowActionToolbar';
-import TopToolbar from '@components/admin/shared/toolbars/TopToolbar';
 import { Typography } from '@components/admin/shared/dom/sanitize';
-import {sanitizeRestProps} from "@admin/utils/props";
-import ErrorOutlineIcon from "@material-ui/icons/ErrorOutline";
-import CheckCircle from "@material-ui/icons/CheckCircle";
-import {fetchJsonWithAuthToken} from "ra-data-django-rest-framework";
+import RowActionToolbar from '@components/admin/shared/toolbars/RowActionToolbar';
+import ShowToolBar from '@components/admin/shared/toolbars/ShowToolbar';
+import TopToolbar from '@components/admin/shared/toolbars/TopToolbar';
 
 const useRaBoxStyles = makeStyles((theme) => ({
   root: {
@@ -96,25 +103,44 @@ const useRaBoxStyles = makeStyles((theme) => ({
 const useArchivesStyles = makeStyles((theme) => ({
   root: {
     display: 'inline-block',
-    marginLeft: theme.spacing(2)
-  }
+    marginLeft: theme.spacing(2),
+  },
 }));
 
 export const PatientShow = (props) => {
+  return (
+    <Show actions={<TopToolbar showExport={true} {...props} />}>
+      <PatientShowRecord {...props}></PatientShowRecord>
+    </Show>
+  );
+};
+
+export const PatientShowRecord = (props) => {
+  const record = useRecordContext();
+  if (!record) return null;
+  return (
+    <BoxedShowLayout>
+      <PatientShowLayout record={record} {...props}></PatientShowLayout>
+    </BoxedShowLayout>
+  );
+};
+
+export const PatientShowLayout = ({ record, props }) => {
   const t = useTranslate();
   const locale = useLocale();
   const notify = useNotify();
+  const resource = useResourceContext();
+  const { permissions } = usePermissions();
   const classes = useRaBoxStyles();
   const archivesToggleClasses = useArchivesStyles();
   const [sort, setSort] = useState({ field: 'start_date', order: 'DESC' });
   const [openDialogEmail, setOpenDialogEmail] = useState(false);
   const [archives, setArchives] = useState(false);
-  const {data, ids, total, loaded} = useGetList(
-    'plans',
-    { page: 1, perPage: 9999},
-    { field: 'start_date', order: 'DESC' },
-    { language: locale, patient_id: props.id, archived: archives}
-  );
+  const { data, total, isLoading } = useGetList('plans', {
+    pagination: { page: 1, perPage: 9999 },
+    sort: { field: 'start_date', order: 'DESC' },
+    filter: { language: locale, patient_id: record.id, archived: archives },
+  });
   const handleArchivesChange = (event) => {
     setArchives(event.target.checked);
   };
@@ -129,47 +155,44 @@ export const PatientShow = (props) => {
 
   const handleSendOnboarding = (event) => {
     setOpenDialogEmail(false);
-    const url = `${process.env.API_ENDPOINT}/patients/${props.id}/resend/`;
+    const url = `${process.env.API_ENDPOINT}/patients/${record.id}/resend/`;
     fetchJsonWithAuthToken(url, {
       method: 'POST',
-      body: JSON.stringify({'confirm': true})
+      body: JSON.stringify({ 'confirm': true }),
     })
-    .then(() => {
-      notify('resources.patients.notifications.email.send.success', 'info');
-    })
-    .catch((e) => {
-      notify('resources.patients.notifications.email.send.failure', 'error');
-    })
-    .finally(() => {
-    });
+      .then(() => {
+        notify('resources.patients.notifications.email.send.success', {
+          type: 'info',
+        });
+      })
+      .catch((e) => {
+        notify('resources.patients.notifications.email.send.failure', {
+          type: 'error',
+        });
+      })
+      .finally(() => {});
   };
 
   return (
-    <Show
-      actions={<TopToolbar showExport={true}/>}
-      {...props}
-    >
-      <BoxedShowLayout>
-        <RaBox className={classes.root}>
-          <RaBox className={classes.leftColumn}>
-            <Typography variant="h6" gutterBottom>
-              {t('admin.shared.labels.card.identity')}
-            </Typography>
-            <RaBox className={classes.columnChild}>
-              <RaBox className={classes.innerChild}>
-                <TextField source="first_name" />
-                <div>
+    <>
+      <RaBox className={classes.root}>
+        <RaBox className={classes.leftColumn}>
+          <Typography variant="h6" gutterBottom>
+            {t('admin.shared.labels.card.identity')}
+          </Typography>
+          <RaBox className={classes.columnChild}>
+            <RaBox className={classes.innerChild}>
+              <TextField source="first_name" />
+              <div>
                 <EmailField source="email" />
                 <IconButton
                   size="small"
-                  style={{marginLeft: '0.5em'}}
+                  style={{ marginLeft: '0.5em' }}
                   onClick={handleOpenDialogEmail}
                 >
                   <MailOutlineIcon />
                 </IconButton>
-                <Dialog
-                  open={openDialogEmail}
-                >
+                <Dialog open={openDialogEmail}>
                   <DialogTitle>
                     {t('admin.shared.text.emailDialog.title')}
                   </DialogTitle>
@@ -199,118 +222,112 @@ export const PatientShow = (props) => {
                     </Button>
                   </DialogActions>
                 </Dialog>
-                </div>
-              </RaBox>
-              <RaBox className={classes.innerChild}>
-                <TextField source="last_name" />
-              </RaBox>
+              </div>
             </RaBox>
-          </RaBox>
-
-          <RaBox className={classes.rightColumn}>
-            <Typography variant="h6" gutterBottom>
-              {t('admin.shared.labels.card.informations')}
-            </Typography>
-            <RaBox className={classes.columnChild}>
-              <RaBox className={classes.innerChild}>
-                <FunctionField
-                  label={t('resources.patients.fields.use_audio')}
-                  render={(record) =>
-                    t(`resources.patients.shared.audio.${record.use_audio}`)
-                  }
-                />
-                <FunctionField
-                  label={t('resources.patients.fields.language')}
-                  render={(record) => t(`languages.${record.language}`)}
-                />
-              </RaBox>
-              <RaBox className={classes.innerChild}>
-                <FunctionField
-                  label={t('resources.patients.fields.side')}
-                  render={(record) =>
-                    t(`resources.patients.shared.side.${record.side}`)
-                  }
-                />
-                <ClinicianTextField show={props.permissions === 'admin'} />
-              </RaBox>
+            <RaBox className={classes.innerChild}>
+              <TextField source="last_name" />
             </RaBox>
           </RaBox>
         </RaBox>
 
-        <Typography variant="h6" gutterBottom gutterTop={true}>
-          {t('resources.patients.card.labels.plans')}
-          <FormGroup classes={archivesToggleClasses}>
-            <FormControlLabel
-              control={
-                <Switch
-                  checked={archives}
-                  onChange={handleArchivesChange}
-                  color="primary"
-                />
-              }
-              label={t('resources.plans.fields.archived')}
-            />
-          </FormGroup>
-          <AddPlanButton patientUid={props.id} />
-        </Typography>
-        <div>
-          {!loaded && <Spinner />}
-          {loaded && (
-            <ListContextProvider
-              value={{
-                data: data,
-                ids: ids,
-                total,
-                currentSort: sort,
-                basePath: props.basePath,
-                resource: props.resource,
-                selectedIds: [],
-              }}
-              {...sanitizeRestProps(props, ['label'],true)}
-            >
-              <Datagrid>
-                <TextField
-                  label={t('resources.plans.fields.name')}
-                  source={`i18n.name.${locale}`}
-                />
-                <NumberField
-                  textAlign="center"
-                  label={t('resources.plans.fields.daily_repeat')}
-                  source="daily_repeat"
-                />
-                <FunctionField
-                  textAlign="center"
-                  label={t('resources.plans.list.labels.exercises_count')}
-                  render={(record) => record.exercises.length}
-                />
-                <DateField
-                  textAlign="center"
-                  label={t('resources.plans.fields.start_date')}
-                  source="start_date"
-                />
-                <DateField
-                  textAlign="center"
-                  label={t('resources.plans.fields.end_date')}
-                  source="end_date"
-                />
-                <RowActionToolbar
-                  basePath='/plans'
-                  rowResource='plans'
-                  context={{patientUid: props.id}}
-                  activable={true}
-                  clonable={true}
-                />
-              </Datagrid>
-            </ListContextProvider>
-          )}
-        </div>
+        <RaBox className={classes.rightColumn}>
+          <Typography variant="h6" gutterBottom>
+            {t('admin.shared.labels.card.informations')}
+          </Typography>
+          <RaBox className={classes.columnChild}>
+            <RaBox className={classes.innerChild}>
+              <FunctionField
+                label={t('resources.patients.fields.use_audio')}
+                render={(record) =>
+                  t(`resources.patients.shared.audio.${record.use_audio}`)
+                }
+              />
+              <FunctionField
+                label={t('resources.patients.fields.language')}
+                render={(record) => t(`languages.${record.language}`)}
+              />
+            </RaBox>
+            <RaBox className={classes.innerChild}>
+              <FunctionField
+                label={t('resources.patients.fields.side')}
+                render={(record) =>
+                  t(`resources.patients.shared.side.${record.side}`)
+                }
+              />
+              <ClinicianTextField show={permissions === 'admin'} />
+            </RaBox>
+          </RaBox>
+        </RaBox>
+      </RaBox>
 
-        <ShowToolBar
-          resource={props.resource}
-          record={props.record}
-          basePath={props.basePath}
-        />
-      </BoxedShowLayout>
-    </Show>
+      <Typography variant="h6" gutterBottom gutterTop={true}>
+        {t('resources.patients.card.labels.plans')}
+        <FormGroup classes={archivesToggleClasses}>
+          <FormControlLabel
+            control={
+              <Switch
+                checked={archives}
+                onChange={handleArchivesChange}
+                color="primary"
+              />
+            }
+            label={t('resources.plans.fields.archived')}
+          />
+        </FormGroup>
+        <AddPlanButton patientUid={record.id} />
+      </Typography>
+      <div>
+        {isLoading && <Spinner />}
+        {!isLoading && (
+          <ListContextProvider
+            value={{
+              data: data,
+              total,
+              sort: sort,
+              resource: resource,
+              selectedIds: [],
+            }}
+            {...sanitizeRestProps(props, ['label'], true)}
+          >
+            <Datagrid bulkActionButtons={false}>
+              <TextField
+                label={t('resources.plans.fields.name')}
+                source={`i18n.name.${locale}`}
+              />
+              <NumberField
+                textAlign="center"
+                label={t('resources.plans.fields.daily_repeat')}
+                source="daily_repeat"
+              />
+              <FunctionField
+                textAlign="center"
+                label={t('resources.plans.list.labels.exercises_count')}
+                render={(record) => record.exercises.length}
+              />
+              <DateField
+                textAlign="center"
+                label={t('resources.plans.fields.start_date')}
+                source="start_date"
+              />
+              <DateField
+                textAlign="center"
+                label={t('resources.plans.fields.end_date')}
+                source="end_date"
+              />
+              <RowActionToolbar
+                basePath="/plans"
+                rowResource="plans"
+                record={record}
+                context={{ patientUid: record.id }}
+                activable={true}
+                clonable={true}
+              />
+            </Datagrid>
+          </ListContextProvider>
+        )}
+      </div>
+
+      <ShowToolBar basePath="/patients" />
+    </>
   );
 };
