@@ -19,59 +19,63 @@
  * You should have received a copy of the GNU General Public License
  * along with MEPP.  If not, see <http://www.gnu.org/licenses/>.
  */
-
 import React, { useCallback, useEffect, useState } from 'react';
 import {
   ArrayInput,
   Create,
+  FormDataConsumer,
   NumberInput,
   SimpleForm,
   SimpleFormIterator,
   TextInput,
   TranslatableInputs,
   useLocale,
+  usePermissions,
+  useRecordContext,
   useTranslate,
 } from 'react-admin';
-import { Typography } from '@components/admin/shared/dom/sanitize';
 
-import {requiredLocalizedField} from '@components/admin/shared/validators';
+import ExerciseRow from '@components/admin/plans/ExerciseRow';
+import IsSystemInput from '@components/admin/plans/IsSystem';
+import { contextualRedirect, preSave } from '@components/admin/plans/callbacks';
 import {
   useSimpleFormIteratorStyles,
   useTranslatorInputStyles,
 } from '@components/admin/plans/styles';
 import { validateExercises } from '@components/admin/plans/validators';
-import { validateNumber } from '@components/admin/shared/validators';
-import IsSystemInput from '@components/admin/plans/IsSystem';
-import ExerciseRow from '@components/admin/plans/ExerciseRow';
-import { LANGUAGES } from '../../../locales';
-import SimpleFormToolBar from '@components/admin/shared/toolbars/SimpleFormToolBar';
+import { Typography } from '@components/admin/shared/dom/sanitize';
 import {
   useGetCategories,
   useGetSubCategories,
 } from '@components/admin/shared/hook';
-import {
-  contextualRedirect,
-  preSave
-} from '@components/admin/plans/callbacks';
+import SimpleFormToolBar from '@components/admin/shared/toolbars/SimpleFormToolbar';
+import { requiredLocalizedField } from '@components/admin/shared/validators';
+import { validateNumber } from '@components/admin/shared/validators';
+
+import { LANGUAGES } from '../../../locales';
 
 export const PlanCreate = (props) => {
+  const record = useRecordContext();
   const t = useTranslate();
+  const { permissions } = usePermissions();
   const locale = useLocale();
   const simpleFormIteratorclasses = useSimpleFormIteratorStyles();
   const translatorClasses = useTranslatorInputStyles();
   const [patientUid, setPatientUid] = useState(undefined);
   const [asTemplate, setAsTemplate] = useState(true);
-  const validateI18n = (record) => {
-    return requiredLocalizedField(record, locale, ['name', 'description']);
-  }
+  const validateI18n = (value, record) => {
+    return requiredLocalizedField(value, record, locale, 'description');
+  };
   const categories = useGetCategories(locale);
   const subCategories = useGetSubCategories(locale);
-  const redirect = useCallback(() => (
-    contextualRedirect(patientUid)
-  ), [patientUid]);
-  const transform = useCallback((record) => (
-    preSave(record, locale, patientUid, asTemplate)
-  ), [patientUid, asTemplate]);
+  const redirect = useCallback(
+    () => contextualRedirect(patientUid),
+    [patientUid],
+  );
+  const transform = useCallback(
+    (record) => preSave(record, locale, patientUid, asTemplate),
+    [patientUid, asTemplate],
+  );
 
   useEffect(() => {
     setPatientUid(props?.history?.location?.state?.patientUid);
@@ -85,13 +89,10 @@ export const PlanCreate = (props) => {
   }, [patientUid]);
 
   return (
-    <Create
-      {...props}
-    >
+    <Create {...props}>
       <SimpleForm
-        validate={validateI18n}
         redirect={redirect}
-        toolbar={<SimpleFormToolBar identity={false} transform={transform}/>}
+        toolbar={<SimpleFormToolBar identity={false} transform={transform} />}
       >
         <Typography variant="h6" gutterBottom>
           {t('resources.plans.card.labels.definition')}
@@ -101,23 +102,20 @@ export const PlanCreate = (props) => {
           defaultLocale={locale}
           classes={translatorClasses}
         >
-          <TextInput
-            source="i18n.name"
-            fullWidth
-          />
+          <TextInput source="i18n.name" validate={validateI18n} fullWidth />
           <TextInput
             source="i18n.description"
+            validate={validateI18n}
             fullWidth
             multiline
           />
         </TranslatableInputs>
-        {props.permissions === 'admin' && asTemplate &&
-          <IsSystemInput/>
-        }
-        <NumberInput
-          source="daily_repeat"
-          validate={validateNumber}
-        />
+        {permissions === 'admin' && asTemplate && (
+          <FormDataConsumer>
+            {({ formData, ...rest }) => <IsSystemInput data={formData} />}
+          </FormDataConsumer>
+        )}
+        <NumberInput source="daily_repeat" validate={validateNumber} />
 
         <Typography variant="h6" gutterBottom gutterTop={true}>
           {t('resources.plans.card.labels.exercises')}
@@ -129,9 +127,7 @@ export const PlanCreate = (props) => {
           label=""
           validate={validateExercises}
         >
-          <SimpleFormIterator
-            classes={simpleFormIteratorclasses}
-          >
+          <SimpleFormIterator classes={simpleFormIteratorclasses}>
             <ExerciseRow
               categories={categories}
               subCategories={subCategories}
