@@ -19,43 +19,48 @@
  * You should have received a copy of the GNU General Public License
  * along with MEPP.  If not, see <http://www.gnu.org/licenses/>.
  */
-
-import React, {useEffect, useState} from 'react';
+import { BoxedShowLayout } from 'ra-compact-ui';
+import React, { useEffect, useState } from 'react';
 import {
-  ArrayField, BooleanField,
+  ArrayField,
+  BooleanField,
   Datagrid,
-  NumberField, NumberInput,
+  NumberField,
+  NumberInput,
   Show,
   TextField,
   TranslatableFields,
   useLocale,
+  usePermissions,
   useTranslate,
 } from 'react-admin';
-import { BoxedShowLayout } from 'ra-compact-ui';
+
+import { makeStyles } from '@mui/styles';
+
+import ClinicianTextField from '@components/admin/clinicians/ClinicianTextField';
 import { useTranslatorInputStyles } from '@components/admin/exercises/styles';
+import { Typography } from '@components/admin/shared/dom/sanitize';
+import { useOnelineStyles } from '@components/admin/shared/styles/oneline';
 import ShowToolBar from '@components/admin/shared/toolbars/ShowToolbar';
 import TopToolbar from '@components/admin/shared/toolbars/TopToolbar';
-import {LANGUAGES} from "../../../locales";
-import { Typography } from '@components/admin/shared/dom/sanitize';
-import ClinicianTextField
-  from '@components/admin/clinicians/ClinicianTextField';
-import { makeStyles } from '@material-ui/core/styles';
-import { useOnelineStyles } from '@components/admin/shared/styles/oneline';
+
+import { LANGUAGES } from '../../../locales';
 
 const useStyles = makeStyles((theme) => {
   return {
     root: {
       '& .ra-field-exercises': {
         marginTop: '-15px',
-        '& > .MuiFormControl-root' : {
-          width: '100%'
-        }
-      }
-    }
+        '& > .MuiFormControl-root': {
+          width: '100%',
+        },
+      },
+    },
   };
 });
 
 export const PlanShow = (props) => {
+  const { permissions } = usePermissions();
   const t = useTranslate();
   const locale = useLocale();
   const translatorClasses = useTranslatorInputStyles();
@@ -79,21 +84,15 @@ export const PlanShow = (props) => {
         <Typography variant="h6" gutterBottom>
           {t('resources.plans.card.labels.definition')}
         </Typography>
-        <ClinicianTextField show={props.permissions === 'admin'} />
+        <ClinicianTextField show={permissions === 'admin'} />
 
         <TranslatableFields
           locales={LANGUAGES}
           defaultLocale={locale}
           classes={translatorClasses}
         >
-          <TextField
-            source="i18n.name"
-            fullWidth={true}
-          />
-          <TextField
-            source="i18n.description"
-            fullWidth={true}
-          />
+          <TextField source="i18n.name" fullWidth={true} />
+          <TextField source="i18n.description" fullWidth={true} />
         </TranslatableFields>
         <NumberField source="daily_repeat" className={onelineClasses.oneline} />
         <BooleanField source="is_system" className={onelineClasses.oneline} />
@@ -102,7 +101,7 @@ export const PlanShow = (props) => {
         </Typography>
 
         <ArrayField source="exercises" label="">
-          <Datagrid>
+          <Datagrid bulkActionButtons={false}>
             <TextField
               source={`i18n.description.${locale}`}
               label={t('resources.plans.fields.exercise.description')}
@@ -122,12 +121,7 @@ export const PlanShow = (props) => {
           </Datagrid>
         </ArrayField>
 
-        <ShowToolBar
-          resource={props.resource}
-          record={props.record}
-          basePath={props.basePath}
-          patientUid={patientUid}
-        />
+        <ShowToolBar basePath="/plans" patientUid={patientUid} />
       </BoxedShowLayout>
     </Show>
   );
