@@ -19,31 +19,34 @@
  * You should have received a copy of the GNU General Public License
  * along with MEPP.  If not, see <http://www.gnu.org/licenses/>.
  */
-
-import React, { useMemo } from 'react';
-import Button from '@material-ui/core/Button';
-import ContentCreate from '@material-ui/icons/Create';
-import ImageEye from '@material-ui/icons/RemoveRedEye';
-
-import { Link } from 'react-router-dom';
 import { linkToRecord, useResourceContext } from 'ra-core';
+import React, { useMemo } from 'react';
+import { Link } from 'react-router-dom';
+
+import ContentCreate from '@mui/icons-material/Create';
+import ImageEye from '@mui/icons-material/RemoveRedEye';
+import Button from '@mui/material/Button';
+
 import { sanitizeRestProps } from '@admin/utils/props';
 
 // useful to prevent click bubbling in a datagrid with rowClick
-const stopPropagation = e => e.stopPropagation();
+const stopPropagation = (e) => e.stopPropagation();
 
 const CRUDButton = (props) => {
   const {
-      basePath = '',
-      label = false,
-      record,
-      scrollToTop = true,
-      type = 'edit',
-      color = 'primary',
-      variant = undefined,
-      ...rest
+    basePath = '',
+    label = false,
+    resource,
+    record,
+    scrollToTop = true,
+    type = 'edit',
+    color = 'primary',
+    variant = undefined,
+    ...rest
   } = props;
-  const resource = useResourceContext();
+  if (!props.resource) {
+    const resource = useResourceContext();
+  }
 
   let icon = <ContentCreate />;
   let endpoint = '';
@@ -57,39 +60,40 @@ const CRUDButton = (props) => {
       break;
   }
   const state = { _scrollToTop: scrollToTop };
-  const context = rest.hasOwnProperty('context') && rest.context
-    ? {...state, ...props.context}
-    : state;
+  const context =
+    rest.hasOwnProperty('context') && rest.context
+      ? { ...state, ...props.context }
+      : state;
   const handleClick = rest.handleClick || stopPropagation;
   const location = rest.hasOwnProperty('location')
     ? props.location
     : record
-      ? `${linkToRecord(basePath || `/${resource}`, record.id)}${endpoint}`
-      : '';
+    ? `${linkToRecord(
+        basePath || `/${props.resource || resource}`,
+        record.id,
+      )}${endpoint}`
+    : '';
 
   return (
     <Button
       component={Link}
-      to={useMemo(() => ({
+      to={useMemo(
+        () => ({
           pathname: location,
           state: context,
         }),
-        [location, context]
+        [location, context],
       )}
       onClick={handleClick}
       color={color}
       variant={variant}
       startIcon={label !== false && icon}
-      {...sanitizeRestProps(rest, [
-        'context',
-        'redirectToBasePath'
-      ], true)}
+      {...sanitizeRestProps(rest, ['context', 'redirectToBasePath'], true)}
     >
       {label !== false && label}
       {label === false && icon}
     </Button>
   );
 };
-
 
 export default CRUDButton;
